requirements.txt
openai==1.3.0
flask==3.0.0
PyYAML==6.0.1
pandas==2.1.3
numpy==1.24.3
jinja2==3.1.2
python-multipart==0.0.6
